requirements.txt
astroid==2.2.5
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
Django==2.2.1
django-cors-headers==3.0.0
django-filter==2.1.0
djangorestframework==3.9.4
googlemaps==3.0.2
idna==2.8
isort==4.3.19
lazy-object-proxy==1.4.1
Markdown==3.1
mccabe==0.6.1
pylint==2.3.1
pytz==2019.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.5
urllib3==1.24.3
wrapt==1.11.1
